requirements.txt
requests>=2.31.0          # For GitHub API calls
python-dotenv>=1.0.0      # For environment variables
mcp[cli]>=1.4.0          # For MCP server functionality
notion-client>=2.3.0      # For Notion integration
langgraph>=0.4.7          # For agentic functionality
langchain>=0.3.25         # For LangChain core functionality
langchain-openai>=0.3.18  # For OpenAI integration
langchain-mcp-adapters>=0.1.1 # For MCP integration
openai>=1.12.0           # For OpenAI embeddings

